LayoutTests/fast/dom/Window/open-invalid-url-expected.txt

CONSOLE MESSAGE: line 1: Unable to open a window with invalid URL '/'.

ALERT: PASS
